Moved most od stuff from main.py to separate files
