+++ src/fun.py
@@ -1,0 +1,2 @@
+Hell = ValueError
+Gayming = ZeroDivisionError
