+++ src/ui.py
@@ -1,0 +1,8 @@
+import pygame, sys
+from pygame.locals import *
+
+
+class fonts:
+    Default = pygame.font.SysFont(None, 24)
+    def __init__(self):
+        pass
